# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SIDAK (Sistem Data Kewilayahan) API
  version: 1.0.0
  description: >
    Spesifikasi Single Source of Truth (SSOT) API untuk sistem pelayanan administrasi
    kelurahan dan portal CMS publik SIDAK. Mendukung sinkronisasi offline-first mobile,
    verifikasi berkas, render PDF asinkron, penyimpanan berkas lokal, dan manajemen konten profil wilayah.
servers:
  - url: /api/v1
    description: Server API Utama
...
paths:
...
  /auth/login:
    post:
      tags: [Autentikasi]
      summary: Login pengguna menggunakan NIK (16 digit) atau NIP (18 digit)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          headers:
            Set-Cookie:
              schema:
                type: string
                example: access_token=eyJhbG...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Autentikasi]
      summary: Logout dan pembersihan cookie sesi
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Logout berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
  /auth/me:
    get:
      tags: [Autentikasi]
      summary: Mendapatkan profil pengguna yang sedang aktif
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================================================
  # 2. PENYIMPANAN BERKAS LOKAL (STORAGE)
  # ============================================================================
  /storage/upload:
    post:
      tags: [Penyimpanan Berkas]
      summary: Mengunggah berkas lampiran atau media ke penyimpanan lokal server
      security:
        - CookieAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, category]
              properties:
                category:
                  type: string
                  enum: [lampiran, cms, dokumen]
                  example: "lampiran"
                transaksi_id:
                  type: string
                  description: ULID transaksi (wajib jika category=lampiran)
                  example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Berkas berhasil diunggah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================================================
  # 3. SINKRONISASI MOBILE (OFFLINE-FIRST)
  # ============================================================================
  /sync/commit:
    post:
      tags: [Sinkronisasi Mobile]
      summary: Commit transaksi form beserta path referensi berkas lampiran
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncCommitRequest'
      responses:
        '201':
          description: Transaksi berhasil disimpan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
        '400':
          description: Validasi JSON Schema gagal atau terdeteksi duplikasi logis 24 jam
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Transaksi duplikat (sudah pernah diproses sebelumnya)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
        '429':
          description: Request dengan Idempotency-Key ini sedang diproses sistem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # 4. PELAYANAN & VERIFIKASI (ADMIN & BACKEND)
  # ============================================================================
  /template-form:
    get:
      tags: [Pelayanan & Verifikasi]
      summary: Mengambil daftar skema form pelayanan aktif
      security:
        - CookieAuth: []
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateForm'
  /transaksi:
    get:
      tags: [Pelayanan & Verifikasi]
      summary: Mengambil daftar antrean transaksi pelayanan
      security:
        - CookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [menunggu_review, sudah_di_review, butuh_revisi]
        - name: layanan_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransaksiListResponse'
  /transaksi/{id}:
    get:
      tags: [Pelayanan & Verifikasi]
      summary: Mengambil detail permohonan beserta link lampiran
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransaksiDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /transaksi/{id}/review:
    patch:
      tags: [Pelayanan & Verifikasi]
      summary: Memperbarui status verifikasi transaksi oleh Lurah/Seklur/Kasi
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewTransaksiRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'

  # ============================================================================
  # 5. DOKUMEN & PDF ENGINE
  # ============================================================================
  /layanan/{id}/generate-pdf:
    post:
      tags: [Dokumen & PDF]
      summary: Memicu antrean asinkron render PDF via Gotenberg
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Permintaan render PDF diterima dan sedang diproses antrean
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    example: "PROCESSING"
        '400':
          $ref: '#/components/responses/BadRequestError'
  /dokumen/{id}/status:
    get:
      tags: [Dokumen & PDF]
      summary: Polling status pembuatan PDF & mendapatkan URL unduhan
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DokumenStatusResponse'

  # ============================================================================
  # 6. CMS PUBLIK (UNAUTHENTICATED - ISR CACHED)
  # ============================================================================
...
  /public/konten:
    get:
      tags: [CMS Publik]
      summary: Mengambil daftar berita dan pengumuman yang telah dipublikasikan
      parameters:
        - name: tipe
          in: query
          schema:
            type: string
            enum: [BERITA, PENGUMUMAN]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KontenPublikListResponse'
  /public/konten/{slug}:
    get:
      tags: [CMS Publik]
      summary: Mengambil detail berita atau pengumuman berdasarkan slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            example: "penyaluran-bansos-tahap-2"
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KontenPublikDetail'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # 7. CMS ADMIN (AUTHENTICATED LURAH/SEKLUR/KASI)
  # ============================================================================
  /cms/profil:
    put:
      tags: [CMS Admin]
      summary: Memperbarui data profil wilayah singleton
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilWilayahInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
  /cms/menu:
    get:
      tags: [CMS Admin]
      summary: Mengambil seluruh menu (termasuk yang nonaktif)
      security:
        - CookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NavigasiMenuAdmin'
    post:
      tags: [CMS Admin]
      summary: Membuat item navigasi menu baru
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigasiMenuInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
  /cms/menu/{id}:
    put:
      tags: [CMS Admin]
      summary: Memperbarui item navigasi menu
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigasiMenuInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
    delete:
      tags: [CMS Admin]
      summary: Menghapus item navigasi menu
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
  /cms/konten:
    get:
      tags: [CMS Admin]
      summary: Mengambil daftar konten untuk panel admin
      security:
        - CookieAuth: []
      parameters:
        - name: tipe
          in: query
          schema:
            type: string
            enum: [BERITA, PENGUMUMAN]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KontenPublikAdminListResponse'
    post:
      tags: [CMS Admin]
      summary: Membuat berita atau pengumuman baru
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KontenPublikInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
  /cms/konten/{id}:
    put:
      tags: [CMS Admin]
      summary: Memperbarui isi berita atau status publikasi
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KontenPublikInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'
    delete:
      tags: [CMS Admin]
      summary: Menghapus berita atau pengumuman
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardMessageResponse'

# ==============================================================================
# KOMPONEN & SKEMA REUSABLE
# ==============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: ULID unik yang digenerate oleh klien per transaksi untuk mencegah duplikasi
      schema:
        type: string
        minLength: 26
        maxLength: 26
        example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
  responses:
    BadRequestError:
      description: Parameter atau payload permintaan tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Sesi tidak sah atau token telah kedaluwarsa
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Data atau sumber daya yang diminta tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    StandardMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Operasi berhasil dieksekusi"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Payload data isian tidak sesuai JSON Schema layanan"
        details:
          type: array
          items:
            type: string

    # Auth Schemas
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: NIK (16 digit) untuk Kader atau NIP (18 digit) untuk Pegawai/Lurah/Seklur/Kasi
          example: "198001012005011001"
        password:
          type: string
          format: password
          example: "Rahasia123!"
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT Access Token untuk aplikasi mobile
          example: "eyJhbGciOiJIUzI1NiIsIn..."
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      required: [id, nama, role]
      properties:
        id:
          type: string
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        nik:
          type: string
          nullable: true
          example: "3205012345670001"
        nip:
          type: string
          nullable: true
          example: "198001012005011001"
        nama:
          type: string
          example: "Asep Sunandar"
        email:
          type: string
          nullable: true
          example: "asep@kelurahan.go.id"
        role:
          type: string
          enum: [LURAH, SEKLUR, KASI, KADER]
          example: "KADER"

    # Storage Schemas
    FileUploadResponse:
      type: object
      required: [file_path, file_url]
      properties:
        file_path:
          type: string
          description: Path relatif penyimpanan lokal
          example: "uploads/lampiran/3205019876540002/01ARZ_ktp.jpg"
        file_url:
          type: string
          description: URL publik untuk akses berkas
          example: "http://localhost:8080/uploads/lampiran/3205019876540002/01ARZ_ktp.jpg"

    # Sync Schemas
    SyncCommitRequest:
      type: object
      required: [transaksi_id, warga_nik, layanan_id, data_isian]
      properties:
        transaksi_id:
          type: string
          minLength: 26
          maxLength: 26
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        warga_nik:
          type: string
          minLength: 16
          maxLength: 16
          example: "3205019876540002"
        layanan_id:
          type: string
          example: "SKTM"
        data_isian:
          type: object
          description: Dynamic payload divalidasi via template_form skema_json
          example:
            keperluan: "Pendaftaran Sekolah"
            penghasilan_bulanan: 500000
            jumlah_tanggungan: 3
        lampiran:
          type: array
          items:
            type: string
          example: ["uploads/lampiran/3205019876540002/01ARZ_ktp.jpg"]

    # Template Form & Transaksi Pelayanan
    TemplateForm:
      type: object
      required: [layanan_id, nama_layanan, skema_json, is_active]
      properties:
        layanan_id:
          type: string
          example: "SKTM"
        nama_layanan:
          type: string
          example: "Surat Keterangan Tidak Mampu"
        deskripsi:
          type: string
          example: "Form pengajuan SKTM untuk keperluan bantuan atau pendidikan"
        skema_json:
          type: object
          description: JSON Schema draft-07
        is_active:
          type: boolean
          example: true
    TransaksiListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              warga_nik:
                type: string
              layanan_id:
                type: string
              status:
                type: string
                enum: [menunggu_review, sudah_di_review, butuh_revisi]
              created_at:
                type: string
                format: date-time
        meta:
          type: object
          properties:
            current_page:
              type: integer
            total_pages:
              type: integer
            total_records:
              type: integer
    TransaksiDetailResponse:
      type: object
      required: [id, warga_nik, layanan_id, data_isian, status, created_at]
      properties:
        id:
          type: string
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        warga_nik:
          type: string
          example: "3205019876540002"
        layanan_id:
          type: string
          example: "SKTM"
        data_isian:
          type: object
        lampiran_urls:
          type: array
          items:
            type: string
          description: URL unduhan berkas lampiran
        status:
          type: string
          enum: [menunggu_review, sudah_di_review, butuh_revisi]
        catatan_review:
          type: string
          nullable: true
        reviewed_by:
          type: string
          nullable: true
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    ReviewTransaksiRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [sudah_di_review, butuh_revisi]
          example: "sudah_di_review"
        catatan_review:
          type: string
          example: "Dokumen lengkap dan terverifikasi sesuai data lapangan."
    DokumenStatusResponse:
      type: object
      required: [dokumen_id, status]
      properties:
        dokumen_id:
          type: string
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        status:
          type: string
          enum: [PROCESSING, READY, FAILED]
          example: "READY"
        download_url:
          type: string
          nullable: true
          description: URL unduhan file PDF jika status READY

    # Profil Wilayah Schemas
...
    ProfilWilayahInput:
      type: object
      required: [nama_kelurahan, kecamatan, kabupaten_kota, visi, misi, alamat_kantor]
      properties:
        nama_kelurahan:
          type: string
        kecamatan:
          type: string
        kabupaten_kota:
          type: string
        visi:
          type: string
        misi:
          type: array
          items:
            type: string
        sejarah:
          type: string
        alamat_kantor:
          type: string
        kontak_telepon:
          type: string
        kontak_email:
          type: string
        struktur_organisasi_file_path:
          type: string
          nullable: true

    # Navigasi Menu Schemas
...
    NavigasiMenuAdmin:
      type: object
      required: [id, label, url, urutan, is_active]
      properties:
        id:
          type: string
        parent_id:
          type: string
          nullable: true
        label:
          type: string
        url:
          type: string
        urutan:
          type: integer
        is_active:
          type: boolean
    NavigasiMenuInput:
      type: object
      required: [label, url, urutan]
      properties:
        parent_id:
          type: string
          nullable: true
        label:
          type: string
          example: "Visi & Misi"
        url:
          type: string
          example: "/profil/visi-misi"
        urutan:
          type: integer
          example: 1
        is_active:
          type: boolean
          default: true

    # Konten Publik Schemas
    KontenPublikListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              tipe:
                type: string
                enum: [BERITA, PENGUMUMAN]
              judul:
                type: string
              slug:
                type: string
              ringkasan:
                type: string
              thumbnail_url:
                type: string
                nullable: true
              published_at:
                type: string
                format: date-time
        meta:
          type: object
          properties:
            current_page:
              type: integer
            total_pages:
              type: integer
            total_records:
              type: integer
    KontenPublikDetail:
      type: object
      required: [id, tipe, judul, slug, ringkasan, isi_konten, published_at, author_nama]
      properties:
        id:
          type: string
        tipe:
          type: string
          enum: [BERITA, PENGUMUMAN]
        judul:
          type: string
        slug:
          type: string
        ringkasan:
          type: string
        isi_konten:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        published_at:
          type: string
          format: date-time
        author_nama:
          type: string
    KontenPublikAdminListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              tipe:
                type: string
              judul:
                type: string
              slug:
                type: string
              is_published:
                type: boolean
              published_at:
                type: string
                format: date-time
                nullable: true
              author_nama:
                type: string
        meta:
          type: object
          properties:
            current_page:
              type: integer
            total_pages:
              type: integer
            total_records:
              type: integer
    KontenPublikInput:
      type: object
      required: [tipe, judul, ringkasan, isi_konten]
      properties:
        tipe:
          type: string
          enum: [BERITA, PENGUMUMAN]
          example: "BERITA"
        judul:
          type: string
          example: "Jadwal Pelayanan Keliling Administrasi Kependudukan"
        ringkasan:
          type: string
          example: "Pelayanan jemput bola dokumen kependudukan di RW 04."
        isi_konten:
          type: string
          example: "<p>Diberitahukan kepada seluruh warga RW 04 bahwa pelayanan...</p>"
        thumbnail_file_path:
          type: string
          nullable: true
          example: "uploads/public/cms/2026/jadwal-keliling_thumb.jpg"
        is_published:
          type: boolean
          default: false